+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 # Changelog
 
+## [0.3.0] - 11/11/2023
+
+### Added
+
+- `test_prompt_randomness` feature for improved prompt generation.
+- Plot variable distributions utils for quickly displaying prompt variable value distribution across generated prompts.
+
+### Changed
+
+- Public domain characters now used for famous characters variables.
+- Cleaned up codebase, removing duplicates and unused styles.
+
+### Fixed
+
+- Updated Docker GitHub action and codecov settings.
+
 ## [0.2.0] - 24/10/2023
 
 ### Added
+
 - Introduced safety checker with retry mechanism for NSFW content detection.
 - Added tests for the new safety checker.
 
@@ -13,8 +30,10 @@
 ## [0.1.3] - 23/10/2023
 
 ### Added
+
 - Around x1.5 speedup on inference.
 - Famous characters and new elements to styles.
 
 ### Changed
+
 - Switched to Python 3.10.